Add redux store to index-static
