+++ src/index-static.js
@@ -1,8 +1,10 @@
 import React from 'react';
 import ReactDOMServer from 'react-dom/server';
+import {Provider} from 'react-redux';
 import { createMemoryHistory, RouterContext, match } from 'react-router';
 import routes from './routes-static';
 import WithStylesContext from './WithStylesContext';
+import store from './store';
 
 export default (locals, callback) => {
   const history = createMemoryHistory();
@@ -11,9 +13,11 @@ export default (locals, callback) => {
   match({ routes, location }, (error, redirectLocation, renderProps) => {
     const css = [];
     const appHtml = ReactDOMServer.renderToString(
-      <WithStylesContext onInsertCss={styles => css.push(styles._getCss())}>
-        <RouterContext {...renderProps} />
-      </WithStylesContext>
+      <Provider store={store}>
+        <WithStylesContext onInsertCss={styles => css.push(styles._getCss())}>
+          <RouterContext {...renderProps} />
+        </WithStylesContext>
+      </Provider>
     );
     const template = require('raw!./../dist/index-html-template.ejs');
     const appCss = css.length ? `<style type="text/css">${css.join('')}</style>` : '';
