+++ src/store.js
@@ -1,0 +1,33 @@
+/* global process */
+
+import {createStore} from 'redux';
+import {getLessons, getTags} from './util';
+import {setContext, setFilter, setLessons} from './action_creators';
+import reducer from './reducer';
+
+const iconContext = require.context('lessonSrc/', true, /^\.\/[^\/]*\/logo-black\.png/);
+const playlistContext = require.context('raw!lessonSrc/', true, /^\.\/[^\/]*\/playlists\/[^\/]*\.txt$/);
+const lessonContext = require.context('onlyFrontmatter!lessonSrc/', true,
+  /^\.\/[^\/]*\/[^\/]*\/(?!README\.md$)[^\/]*\.md/);
+const lessons = getLessons(lessonContext);
+
+const initialState = {};
+const isProduction = process.env.NODE_ENV === 'production';
+let store;
+
+if (isProduction) {
+  store = createStore(reducer, initialState);
+} else {
+  //Only use the DevTools extension when in development
+  const devTools = typeof window === 'object' && typeof window.devToolsExtension !== 'undefined' ?
+    window.devToolsExtension() :
+    f => f;
+
+  store = createStore(reducer, initialState, devTools);
+}
+store.dispatch(setContext('iconContext', iconContext));
+store.dispatch(setContext('playlistContext', playlistContext));
+store.dispatch(setLessons(lessons));
+store.dispatch(setFilter(getTags(lessonContext)));
+
+export default store;
